Update typescript absolute paths
