+++ src/__tests__/components/molecules/Header.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import { NHSNotifyHeader } from '@/src/components/molecules/Header/Header';
+import { NHSNotifyHeader } from '@molecules/Header/Header';
 
 describe('Header component', () => {
   it('renders component correctly', () => {
