--- src/types/template-format.types.ts
@@ -1,6 +1,0 @@
-export enum TemplateFormatText {
-  APP = 'APP',
-  EMAIL = 'EMAIL',
-  LETTER = 'LETTER',
-  SMS = 'SMS',
-}
