+++ src/utils/form-actions.ts
@@ -1,7 +1,7 @@
 'use server';
 
 import { randomUUID } from 'node:crypto';
-import { getAmplifyBackendClient } from '@/src/utils/amplify-utils';
+import { getAmplifyBackendClient } from '@utils/amplify-utils';
 
 export async function createSession() {
   const { data } = await getAmplifyBackendClient().models.SessionStorage.create(
